Refactor config modules and add loader tests
