+++ docs/ru/CONFIG.md
@@ -8,7 +8,7 @@
 
 Переменные окружения используют схему `CHEMBL_DA__РАЗДЕЛ__КЛЮЧ`, вложенные ключи
 соединяются двойным подчёркиванием. Для распространённых параметров доступны
-короткие алиасы (см. `_ALIAS_MAP` в `library/config.py`).
+короткие алиасы (см. `_ALIAS_MAP` в `library/config/loader.py`).
 
 ```bash
 # Пример: увеличить лимит UniProt во время smoke-теста
