+++ library/config/__init__.py
@@ -1,0 +1,29 @@
+"""Configuration package exposing public schemas and helpers."""
+
+from __future__ import annotations
+
+from .loader import _mask_secrets, _serialize_paths, build_alias_map, load_config, print_config
+from .models import *  # noqa: F401,F403
+from .runtime import (
+    apply_rate_limiter_settings,
+    crossref_session,
+    ensure_dirs,
+    openalex_session,
+    session_with_retry,
+)
+
+from .models import __all__ as _model_all
+
+__all__ = [
+    *_model_all,
+    "build_alias_map",
+    "load_config",
+    "print_config",
+    "_serialize_paths",
+    "_mask_secrets",
+    "session_with_retry",
+    "openalex_session",
+    "crossref_session",
+    "ensure_dirs",
+    "apply_rate_limiter_settings",
+]
